Pretty Print exception condition handled
